Added input for new players and load game
When the user selects to start a new game; they are instructed to insert a name for players 1 and 2, which are currently stored in strings within that method.
When the user selects to load a saved game; they are instructed to insert the name for the saved file in which they intend to load from. This is also stored within a string within that method for now (both subject to change)