90% working save function
Implemented a saveGame() function in main.cpp as opposed to Save.cpp as we decided for only a single function; having its own separate file is a bit unnecessary (but may change if the function becomes too large). Currently when you start a new game you are to input 4 things; the function (turn/save) and 3 parameters (the 3 parameters are intended for the turn mechanic) but if the player enters "save" as the function, and a file name as the first parameter; the saveGame() function is called. The only current issue with this as of now is that the game is still awaiting the other inputs for both parameters 2 and 3 even though they are not needed for this save function, so that will need to be fixed. I also added a vector that stores all of the turns taken throughout the game; so as the player takes a turn (assuming it is a valid turn; validation coming soon) that is added to the vector. When the game is being saved, we iterate though said vector and add each item of the turns vector to a formatted string that is then added to the ostream output file. This keeps going until either the players wish to exit the game without saving and hit ^D, or save the game which automatically saves the game and returns them to the main menu.