Testing mode begins, moved to load.cpp
Fleshed out load.cpp and load.h to have the correct methods, with the filename passing through via the method in main.cpp. Also began work on the testing mode, by altering the main method to be able to pass in values (such as "-t" and "filename" so if the -t identifier is present then it will pass the following value to the testing mode method. I believe some validation testing will be needed to ensure a user actually inputs a file name to be tested, but that will commence in week 9