Relative paths for saving and loading
When saving the game, it is now saved into the "saves" folder, and automatically given the .save extension to keep some kind of formatting together. Also created a new "tests" folder where our test cases will be stored. When calling the loadGame() function; the game will search the "saves" folder for a game matching the input name. Similarly: when calling the testingMode() function - the game will search the "tests" folder for a game matching the input name.