Global Boolean created, load function progress
A new global boolean 'isLoading' is now created; which will be set to true when the loadGame() method is invoked. When a new game is run, it will check whether isLoading is set to true or not. If it is in fact true, then a simplified version of the game will run. This simplified version should act the same as a regular game when complete, except it will be missing the various std::cout's seen in a regular round of Azul (i.e. "Enter a name for player 1", "===Start Round 1===", etc). While working towards this new simplified mode, existing data structures are being modified to take in new parameters, such as the createPlayers() method now taking in both player names, and the cin/cout takes places in the runGame() method only when isLoading is set to false.