loading mode overhaul?
Added many if(isLoading) statements throughout the gameEngine.cpp file, so if the game is loading then the many couts used for playing a game are disabled and the game plays without them in the background. I also majorly overhauled the playerEntersTurn() method to now take the function and its 3 parameters as input, however the 3 parameters are optional variables. This means if the game is being loaded, then all variables are passed through via the save file (i.e. playerEntersTurn(currentPlayer, function, param1, param2, param3);

But if a game is not being loaded then all that must be called is playerEntersTurn(currentPlayer, function);
and the game will function as it used to with no issues. This was the biggest turnover point for me as i struggled for countless hours today trying to figure out how i can pass through values without disrupting the logic of the game that already existed prior to the loadGame method.

The final steps are to:
1. Pass the values stored in load.cpp through to gameEngine.cpp, potentially through the use of getters and setters so they can be called upon anywhere throughout the game where necessary.
2. add the testing mode, which is just the loadGame method, with a slight change on the end. (want to fully implement loadGame first before i start this).
3. clean up all code
4. TBD