overload function created, loadGame now 99% complete
using an overload function for runGame that now passes through the load object, we can now use load->getPlayer1() to get the name for player 1, and similar getters. Now when loading a game, the custom tileBag is loaded, players are created with the correct names, and the game just saves without making any moves for testing.
Last step is to loop the number of turns for how many elements that are in the turns array, then setting isLoading to false to continue the game.