Split turn input when loading game
When loading a game, the turns are added to a vector line by line when loadGame(fileName) is called. We loop through this vector in gameEngine.cpp, and using std::istringstream, add each word that is separated by a space to its own vector 'splitTurn', then add each of those to a local string to call upon it in a much simpler manner later on.