Testing mode 100% complete
The main thing that needs to be finished is checking the file being loaded is valid. Otherwise it is just commenting code and cleaning up code overall. The testing mode is now thankfully 100% complete thanks to a new bool value in the load constructor called testing, which is defaulted to false. It is only set to true when the testingMode method is called in load.cpp, and at the beginning of the runTurn method we check if there are
a) no more turns to be loaded
b) if testing = true
if both conditions are met, the factories and mosaics are printed, and the game is safely exited using std::_Exit(EXIT_SUCCESS)